+++ trestle/core/catalog/catalog_merger.py
@@ -91,13 +91,13 @@ class CatalogMerger():
                 new_control_handle = copy.deepcopy(control_handle)
                 new_control_handle.control = src
                 # add the control and its handle to the param_dict
-                self._catalog_interface._control_dict[src.id] = new_control_handle
+                self._catalog_interface._control_dict[src.id] = new_control_handle  # type: ignore
 
         # now need to cull any controls that are not in the src catalog
         handled_ids = set(cat_interface._control_dict.keys())
         orig_ids = set(self._catalog_interface._control_dict.keys())
         extra_ids = orig_ids.difference(handled_ids)
-        for extra_id in extra_ids:
+        for extra_id in sorted(extra_ids):
             self._catalog_interface._control_dict.pop(extra_id)
 
         self._catalog_interface.update_catalog_controls()