+++ trestle/core/markdown/base_markdown_node.py
@@ -32,10 +32,10 @@ logger = logging.getLogger(__name__)
 class BaseSectionContent:
     """A content of the node."""
 
-    def __init__(self):
+    def __init__(self) -> None:
         """Initialize section content."""
         self.raw_text = ''
-        self.subnodes_keys = []
+        self.subnodes_keys: List[str] = []
 
     def union(self, node: BaseMarkdownNode) -> None:
         """Unites contents together."""
@@ -54,7 +54,7 @@ class BaseMarkdownNode:
         self.starting_line = starting_line
 
     @classmethod
-    def build_tree_from_markdown(cls, lines: List[str]):
+    def build_tree_from_markdown(cls, lines: List[str]) -> BaseMarkdownNode:
         """Construct a tree out of the given markdown."""
         ob = cls.__new__(cls)
         start_level = ob._get_max_header_lvl(lines)
@@ -70,7 +70,7 @@ class BaseMarkdownNode:
     def get_node_for_key(self, key: str, strict_matching: bool = True) -> Optional[BaseMarkdownNode]:
         """Return a first node for the given key, substring matching is supported. The method is case insensitive."""
         if not strict_matching:
-            if not any([key.lower() in el.lower() for el in self.content.subnodes_keys]):
+            if not any(key.lower() in el.lower() for el in self.content.subnodes_keys):
                 return None
             elif len(as_filtered_list(self.content.subnodes_keys, lambda el: key.lower() in el.lower())) > 1:
                 logger.warning(f'Multiple nodes for {key} were found, only the first one will be returned.')
@@ -99,7 +99,7 @@ class BaseMarkdownNode:
         Returns: List of found markdown nodes
         """
         if not strict_matching:
-            if not any([key in el for el in self.content.subnodes_keys for key in keys]):
+            if not any(key in el for el in self.content.subnodes_keys for key in keys):
                 return []
         elif not set(keys).intersection(self.content.subnodes_keys):
             return []
@@ -256,10 +256,10 @@ class BaseMarkdownNode:
         """
         if key.lower() == node.key.lower() or (not strict_matching and key.lower() in node.key.lower()):
             return node
-        if (not strict_matching and any([key.lower() in el.lower()
-                                         for el in node.content.subnodes_keys])) or (key.lower() in [
-                                             el.lower() for el in node.content.subnodes_keys
-                                         ]):
+        if (not strict_matching and any(key.lower() in el.lower()
+                                        for el in node.content.subnodes_keys)) or (key.lower() in [
+                                            el.lower() for el in node.content.subnodes_keys
+                                        ]):
             for subnode in node.subnodes:
                 matched_node = self._rec_traverse(subnode, key, strict_matching)
                 if matched_node is not None:
@@ -320,7 +320,7 @@ class BaseMarkdownNode:
         for subnode in node.subnodes:
             self._rec_traverse_header_update(subnode, header_map)
 
-    def _get_max_header_lvl(self, lines: List[str]):
+    def _get_max_header_lvl(self, lines: List[str]) -> int:
         """Go through all lines to determine highest header level. Less # means higher."""
         min_lvl = math.inf
         for line in lines: